Add injected sidecars to TaskRun Status for proper cleanup
Injected sidecars are only cleaned up if sidecars are defined in the
task spec because of a check that exits early if the TaskRun does not
include any sidecars. This breaks users that are running in environments
with injected sidecars, but are not defining sidecars in their TaskRuns.

This change will result in the status of injected sidecars being added
to the TaskRun Status so that podconvert.IsSidecarStatusRunning can successfully
track their state and exit stopSidecars early in the case that no
sidecars are injected or defined in the spec to prevent unnecesary
api calls to get pods. Side effects of this change are that
non-user-defined sidecars will show in the TaskRun status.

/kind bug

Resolves #4731